Move MemoryLogger to core so we can use it in storage lib as well
