+++ app/src/main/java/com/stevesoltys/seedvault/repo/Checker.kt
@@ -7,7 +7,7 @@ package com.stevesoltys.seedvault.repo
 
 import androidx.annotation.WorkerThread
 import com.google.protobuf.ByteString
-import com.stevesoltys.seedvault.MemoryLogger
+import org.calyxos.seedvault.core.MemoryLogger
 import com.stevesoltys.seedvault.backend.BackendManager
 import com.stevesoltys.seedvault.crypto.Crypto
 import com.stevesoltys.seedvault.proto.Snapshot
